+++ lib/browser_core.js
@@ -6,7 +6,7 @@ const {
 } = require('./errors');
 const WendigoConfig = require('../config');
 const DomElement = require('./models/dom_element');
-const utils = require('./utils');
+const utils = require('./utils/utils');
 
 const injectionScriptsPath = WendigoConfig.injectionScripts.path;
 const injectionScripts = WendigoConfig.injectionScripts.files;
