+++ lib/mixins/browser_actions.js
@@ -2,7 +2,7 @@
 "use strict";
 
 const DomElement = require('../models/dom_element');
-const utils = require('../utils');
+const utils = require('../utils/utils');
 const {FatalError, QueryError} = require('../errors');
 
 module.exports = function BrowserActionsMixin(s) {
