+++ lib/mixins/browser_wait.js
@@ -1,7 +1,7 @@
 /* global WendigoUtils */
 "use strict";
 
-const utils = require('../utils');
+const utils = require('../utils/utils');
 const {TimeoutError} = require('../errors');
 
 module.exports = function BrowserWaitMixin(s) {
