+++ lib/models/request_mock.js
@@ -5,7 +5,7 @@ const URL = require('url').URL;
 
 const RequestFilter = require('../modules/requests/request_filter');
 const {FatalError, AssertionError, TimeoutError} = require('../errors');
-const utils = require('../utils');
+const utils = require('../utils/utils');
 
 
 module.exports = class RequestMock {
