+++ lib/modules/assertions/browser_assertions.js
@@ -2,7 +2,7 @@
 /* global WendigoUtils */
 "use strict";
 
-const utils = require('../../utils');
+const utils = require('../../utils/utils');
 const elementsAssertionUtils = require('../../utils/assert_elements');
 const assertUtils = require('../../utils/assert_utils');
 const RequestAssertionsFilter = require('../requests/request_assertions_filter');
