+++ lib/modules/assertions/browser_not_assertions.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const utils = require('../../utils');
+const utils = require('../../utils/utils');
 const assertUtils = require('../../utils/assert_utils');
 
 module.exports = class BrowserNotAssertions {
