+++ lib/modules/console/browser_console.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const utils = require('../../utils');
+const utils = require('../../utils/utils');
 const Log = require('../../models/log');
 
 module.exports = class BrowserConsole {
