+++ lib/modules/requests/request_assertions_filter.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const {AssertionError} = require('../../errors');
-const utils = require('../../utils');
+const utils = require('../../utils/utils');
 
 module.exports = class RequestAssertionsFilter extends Promise {
     constructor(executor, requestFilter) {
