+++ lib/modules/requests/request_filter.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const utils = require('../../utils');
+const utils = require('../../utils/utils');
 
 
 function processBody(body) {
