+++ lib/modules/requests/request_mocker.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const URL = require('url').URL;
-const utils = require('../../utils');
+const utils = require('../../utils/utils');
 const RequestMock = require('../../models/request_mock');
